\name{apply-methods}
\docType{methods}
\alias{apply-methods}
\alias{apply,big.matrix-method}
\title{apply() for big.matrix objects.}
\description{
\code{\link{apply}} for \code{\link{big.matrix}} objects.
}
\section{Methods}{
\describe{
\item{apply}{\code{signature(x = "big.matrix")}: \code{apply()} where \code{MARGIN} may only be 1 or 2, but otherwise conforming to what you would expect from \code{apply()}.}
}}
\keyword{methods}
