% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-biganalytics.R
\docType{package}
\name{biganalytics-package}
\alias{biganalytics-package}
\alias{biganalytics}
\title{Utilities for big.matrix objects of package bigmemory}
\description{
Extend the \pkg{bigmemory} package with various analytics.  In addition
to the more obvious summary statistics (see \code{\link{colmean}}, etc...),
\pkg{biganalytics} offers \code{\link{biglm.big.matrix}},
\code{\link{bigglm.big.matrix}}, \code{\link{bigkmeans}},
\code{\link{binit}}, and \code{apply} for 
\code{big.matrix} objects.  Some of the functions may be used with native 
\R objects, as well, providing gains in speed and memory-efficiency.
}
